ci: enable pedantic clippy warnings and fix outstanding lints

+++ oro-common/src/mem/pfa.rs
@@ -17,41 +17,41 @@ pub unsafe trait PageFrameAllocator {
 	/// Allocates a new page frame, returning the physical address of the page frame
 	/// that was allocated. If `None` is returned, the system is out of memory.
 	///
-	fn allocate(&mut self) -> Option<usize>;
+	fn allocate(&mut self) -> Option<u64>;
 
 	/// Frees a page frame.
 	///
 	/// # Panics
 	/// Implementors **must** panic if the passed frame address is not page-aligned.
-	fn free(&mut self, frame: usize);
+	fn free(&mut self, frame: u64);
 
 	/// Gets the number of bytes of memory that are currently in use by the system.
-	fn used_memory(&self) -> usize;
+	fn used_memory(&self) -> u64;
 
 	/// Gets the number of bytes of memory that are currently free and available to the system.
 	/// This does not include unusable memory regions.
 	#[inline]
-	fn free_memory(&self) -> usize {
+	fn free_memory(&self) -> u64 {
 		self.total_usable_memory() - self.used_memory()
 	}
 
 	/// Gets the number of bytes of memory in the system that are unusable.
 	/// This **does not** include bad memory.
-	fn total_unusable_memory(&self) -> usize;
+	fn total_unusable_memory(&self) -> u64;
 
 	/// Gets the number of bytes of "bad" memory in the system.
 	/// This is **not** simply unusable memory, but memory explicitly marked
 	/// as "bad" by the bootloader. Returns `None` if the bootloader does not
 	/// provide this information.
-	fn total_bad_memory(&self) -> Option<usize>;
+	fn total_bad_memory(&self) -> Option<u64>;
 
 	/// Gets the total amount of memory, including usable, unusable, and bad memory.
-	fn total_memory(&self) -> usize {
+	fn total_memory(&self) -> u64 {
 		self.total_usable_memory()
 			+ self.total_unusable_memory()
 			+ self.total_bad_memory().unwrap_or(0)
 	}
 
 	/// Gets the total number of bytes of memory that are usable to the system.
-	fn total_usable_memory(&self) -> usize;
+	fn total_usable_memory(&self) -> u64;
 }
